rebase

+++ nexus/SIs/models.py
@@ -33,19 +33,31 @@ class SIRoleInfo(models.Model):
         help_text="The class that this role is for."
     )
     
+    all_sections = models.BooleanField(
+        default=False,
+        help_text="If this role is for all sections of the class."
+    )
+    
     class Meta:
         unique_together = [
             "position",
             "assigned_class",
         ]
     
+    def __str__(self):
+        return f"{self.position.user.first_name} - {self.assigned_class}"
+    
     def save(self, *args, **kwargs):
         if self.position.position != PositionChoices.SI:
             raise ValueError("This role is not for an SI position.")
         if self.id is not None:
             old_role = SIRoleInfo.objects.get(id=self.id)
+            if old_role.assigned_class == self.assigned_class:
+                return super(SIRoleInfo, self).save(*args, **kwargs)
             SIReccuringShiftInfo.objects.filter(role=old_role).delete()
-        super(SIRoleInfo, self).save(*args, **kwargs)
+        ret = super(SIRoleInfo, self).save(*args, **kwargs)
+        if self.assigned_class is None:
+            return ret
         class_times = ClassTimes.objects.filter(orignal_class=self.assigned_class)
         active_semester = Semester.objects.get_active_semester()
         for class_time in class_times:
@@ -65,6 +77,7 @@ class SIRoleInfo(models.Model):
                 class_time=class_time,
                 recuring_shift=rs,
             )
+        return ret
 
 class SIReccuringShiftInfo(models.Model):
     role = models.ForeignKey(
